Fix paragraph moving logic to avoid stuck at line beginning with number
The checking in is_new_para() should match that of inword(). The current
codes only check isletter in is_new_para() while inword() also checks
number. This causes Meta-P (gotobop) to stuck at line beginning with number.

To reproduce the problem, use em to open the README file and do Meta-P
from the end of the file. And it will stuck at the line '1991.'.

Fix it to also check numberic chars is is_new_para() function.

Signed-off-by: Chen Qi <chen_q07@163.com>